pyproject.toml
[tool.poetry]
name = "poetry-dojo"
version = "0.1.0"
description = ""
authors = ["Gabriel Fior <fior.gbm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
dojo-compass = ">=2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
